requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1
aiohttp==3.11.11
beautifulsoup4==4.12.3
lxml==5.3.0
pyyaml==6.0.2
aiosqlite==0.20.0
aiomqtt==2.3.0
httpx==0.28.1
